record_linkage_service/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
influxdb-client>=1.38.0
